+++ cli/maintainer/src/commands_solido.rs
@@ -917,6 +917,7 @@ pub fn command_deactivate_validator_if_commission_exceeds_max(
     let solido = config.client.get_solido(opts.solido_address())?;
 
     let mut violations = vec![];
+    let mut instructions = vec![];
     for pubkey_entry in solido.validators.entries {
         let validator = pubkey_entry.entry;
         let vote_pubkey = pubkey_entry.pubkey;
@@ -935,13 +936,18 @@ pub fn command_deactivate_validator_if_commission_exceeds_max(
                 validator_vote_account_to_deactivate: vote_pubkey,
             },
         );
-        let signers: Vec<&dyn Signer> = vec![];
-        config.sign_and_send_transaction(&[instruction], &signers)?;
+        instructions.push(instruction);
         violations.push(ValidatorViolationInfo {
             validator_vote_account: vote_pubkey,
             commission,
         });
     }
+
+    let signers: Vec<&dyn Signer> = vec![];
+    // Due to the fact that Solana has a limit on number of instructions in a transaction
+    // this can fall if there would be alot of misbehaved validators each
+    // exceeding `max_commission_percentage`. But it is very improbable scenario.
+    config.sign_and_send_transaction(&instructions, &signers)?;
 
     Ok(DeactivateValidatorIfCommissionExceedsMaxOutput {
         entries: violations,