+++ program/src/process_management.rs
@@ -55,7 +55,7 @@ pub fn process_add_validator(program_id: &Pubkey, accounts_raw: &[AccountInfo]) 
     // Deserialize also checks if the vote account is a valid Solido vote
     // account: The vote account should be owned by the vote program, the
     // withdraw authority should be set to the program_id, and it should
-    // sattisfy commission limit.
+    // sattisfy the commission limit.
     let _partial_vote_state = PartialVoteState::deserialize(
         accounts.validator_vote_account,
         lido.max_commission_percentage,